backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
